[Bug] Fix windows compilation of catalan_number A warning in the compilation process causing the compilation to fail. An implicit cast from int64 to int32 could have cause a lost of information, fixed by adding a an explicit cast

+++ source/catalan_number.cpp
@@ -45,7 +45,8 @@ int find_catalan(int n)
                 = (a_catalan[i] + (a_catalan[j] * a_catalan[i - j - 1]) % 1000000007) % 1000000007;
     }
 
-    int res = a_catalan[n];
+    int res = (int)a_catalan[n];  // It is allowed to explicitly cast to int because the result is
+                                  // guaranteed to fit in 32 bit integer after the modulo operation
     delete[] a_catalan;
     return res;
 }
